Move SinkEventAttributeSet to doxia-sink-api
doxia-sink-api declares Sink methods taking a SinkEventAttributes but
ships no implementation of that interface and no factory for one, so
every caller that wants to pass attributes to a sink has to depend on
doxia-core and import from the .impl package. The class was in package
org.apache.maven.doxia.sink until DOXIA-506 moved it in with the sink
implementations.

Move it, along with its nested Semantics, next to the interface it
implements. It brings no new dependencies with it: it needs only
javax.swing.text.AttributeSet, JDK collections and SinkEventAttributes,
and SinkEventAttributes already extends MutableAttributeSet.

The old class stays behind as a deprecated subclass so that existing
bytecode keeps working. It carries its own nested Semantics, which is
not inherited under the old binary name, and overrides unmodifiable() to
keep the old return type. The protected parser hooks in doxia-core that
take this type keep taking the deprecated one, so nothing on the
extension API changes; japicmp passes with no exclusions.
