+++ doxia-core/src/main/java/org/apache/maven/doxia/util/DoxiaUtils.java
@@ -32,7 +32,7 @@ import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.Locale;
 
-import org.apache.maven.doxia.sink.impl.SinkEventAttributeSet;
+import org.apache.maven.doxia.sink.SinkEventAttributeSet;
 
 /**
  * General Doxia utility methods. The methods in this class should not assume
