+++ doxia-core/src/test/java/org/apache/maven/doxia/macro/toc/TocMacroTest.java
@@ -32,8 +32,8 @@ import org.apache.maven.doxia.markup.Markup;
 import org.apache.maven.doxia.parser.AbstractParserTest;
 import org.apache.maven.doxia.parser.Xhtml5BaseParser;
 import org.apache.maven.doxia.sink.Sink;
+import org.apache.maven.doxia.sink.SinkEventAttributeSet;
 import org.apache.maven.doxia.sink.impl.CreateAnchorsForIndexEntriesFactory;
-import org.apache.maven.doxia.sink.impl.SinkEventAttributeSet;
 import org.apache.maven.doxia.sink.impl.SinkEventElement;
 import org.apache.maven.doxia.sink.impl.SinkEventTestingSink;
 import org.apache.maven.doxia.sink.impl.Xhtml5BaseSink;
